Get train movement sync (mostly) working again
This is by far the most efficient method of syncing train position we've had, only taking up 240 bytes per-car per-tick. It might be able to be optimized further, but bandwidth testing should be done first to see if it's necessary.

The main problem is that it's still mostly relying on velocity (speed) alone to sync position. This means that if a client sets a different speed than the server between ticks, positions can still desync. I initially tried to move the bogie's traveller to the server's span with each packet, but that causes extremely jittery movement on the client.