Implement basic player name tags
Name tags can be toggled in settings, along with whether they show the player's ping. Some improvements could be made to the centering of the name text in relation to the background, and ideally they should render on top of any world objects, but that can all be implemented later.

Closes #4.