Sync money, licenses, garages, and in debt status
Syncing each debt and making them payable by clients will take a fair bit of work due to the systems complexity. For now, the fees screen in the career manager has a note that only the host can pay them, and the host only tells the clients whether they're in debt or not so it can be shown in the licenses screen and (eventually) when taking jobs.

Clients can purchase licenses. When they do so, they ask the server to purchase the licenses, which then validates the request, and relays the new license back to all clients.

Until items are synced, clients cannot unlock garages. However, if the host unlocks a garage, it'll get unlocked for clients as well.

Closes #19