Server browser and lobby server working
Server browser now works, as well as the host game panel.
When a multiplayer game starts, the game registers itself with the lobby server and continues to provide updates while the session is active. When the session deactivates the lobby server is notified to remove the game server.

More work required on GUI